// src/store/promocode/promocode.enum.ts
enum PromocodeActionTypeEnum {
  GET_PROMOCODE_SUCCESS = "GET_PROMOCODE_SUCCESS",
  GET_PROMOCODE_PENDING = "GET_PROMOCODE_PENDING",
  GET_PROMOCODE_FAILED = "GET_PROMOCODE_FAILED",

  GET_PROMOCODE_BY_ID_SUCCESS = "GET_PROMOCODE_BY_ID_SUCCESS",
  GET_PROMOCODE_BY_ID_PENDING = "GET_PROMOCODE_BY_ID_PENDING",
  GET_PROMOCODE_BY_ID_FAILED = "GET_PROMOCODE_BY_ID_FAILED",

  UPDATE_PROMOCODE_SUCCESS = "UPDATE_PROMOCODE_SUCCESS",
  UPDATE_PROMOCODE_PENDING = "UPDATE_PROMOCODE_PENDING",
  UPDATE_PROMOCODE_FAILED = "UPDATE_PROMOCODE_FAILED",

  DELETE_PROMOCODE_SUCCESS = "DELETE_PROMOCODE_SUCCESS",
  DELETE_PROMOCODE_PENDING = "DELETE_PROMOCODE_PENDING",
  DELETE_PROMOCODE_FAILED = "DELETE_PROMOCODE_FAILED",

  ADD_PROMOCODE_SUCCESS = "ADD_PROMOCODE_SUCCESS",
  ADD_PROMOCODE_PENDING = "ADD_PROMOCODE_PENDING",
  ADD_PROMOCODE_FAILED = "ADD_PROMOCODE_FAILED",
}

export default PromocodeActionTypeEnum;